+++ examples/data-sources/tilgangsportalen_entra_group/data-source.tf
@@ -1,0 +1,3 @@
+data "tilgangsportalen_entra_group" "example" {
+  name = "[Ex] group 1"
+}