+++ internal/provider/data_source_list_system_roles.go
@@ -3,10 +3,13 @@ package provider
 import (
 	"context"
 	"fmt"
+
 	"github.com/hashicorp/terraform-plugin-framework/datasource"
 	"github.com/hashicorp/terraform-plugin-framework/datasource/schema"
 	"github.com/hashicorp/terraform-plugin-framework/types"
+
 	"github.com/hashicorp/terraform-plugin-log/tflog"
+
 	"terraform-provider-tilgangsportalen/internal/tilgangsportalapi"
 )
 
@@ -64,23 +67,9 @@ func (d *SystemRolesDataSource) Schema(ctx context.Context, req datasource.Schem
 
 // Configure adds the provider configured client to the resource.
 func (d *SystemRolesDataSource) Configure(ctx context.Context, req datasource.ConfigureRequest, resp *datasource.ConfigureResponse) {
-	// Prevent panic if the provider has not been configured.
-	if req.ProviderData == nil {
-		return
-	}
-
-	client, ok := req.ProviderData.(*tilgangsportalapi.Client)
-
-	if !ok {
-		resp.Diagnostics.AddError(
-			"Unexpected Data Source Configure Type",
-			fmt.Sprintf("Expected *tilgangsportalapi.Client, got: %T. Please report this issue to the provider developers.", req.ProviderData),
-		)
-
-		return
-	}
-
-	d.client = client
+	ConfigureClientDataSource(ctx, req, resp, func(client *tilgangsportalapi.Client) {
+		d.client = client
+	})
 }
 
 // Read calls the API to get the latest data for the resource